Add --with-orte-lib and --with-orte-include to configure
To be able to set library and include path independently from each other
we need the option to define those separately. Whilst --with-orte can
point to a directory that contains lib/ and include/ directory we can
use the new switches to point either lib or include to a seperate
directory.