Improve configuration for the STCI back end
Add --with-stci=/path/to/stci to set the path to the stci configuration.
Add --with-stci-lib & --with stci-include configuration options to set
these paths if they divert from the standard path.

Changed variable names in Makefile.mk & configure to have a consistent
scheme with the variable names used for the opal configuration.