Add back end detection for ORTE.
Rename the function table structure and variable to make it easier not
to exceed 80 columns in the source.

Add detection for the ORTE back end by utilizing an ORTE specific
environment variable. If it is set the ORTE back end will be
initialized, otherwise we initialize STCI. This needs to be extended as
soon as we add more back ends.