Fix some pmi related problems with srs.
The PMI key value store only accepts string values. To be able to store any value we
have to make sure that the string is not terminated by a 0x0 to early.
The current implementation masks byte/multi byte values to make sure we do not get
a 0x0 byte. However, it is not perfect and still needs some improvement.